:recycle: refactor : [Cpp07Ex01] refactoring type

+++ cpp07/ex01/include/iter.hpp
@@ -13,8 +13,8 @@
 #define ITER_HPP
 
 template <typename T>
-void iter(T *array, unsigned long long size, void (*f)(T)) {
-  for (unsigned long long i = 0; i < size; i++) {
+void iter(T *array, size_t size, void (*f)(T)) {
+  for (size_t i = 0; i < size; i++) {
     f(array[i]);
   }
 }
